Make OCPP library compatible with Android 8
Android 8's Jack has issues processing class files with lambda
operations that implicitly use classes which are not part of the
project, resulting in build failures.

Replace the four occurrences where such lambdas are used with
functionally identical but less shorthand code, so that the OCPP
library will build for Android 8.