Allow omitting all optional SampledValue fields
According to the OCPP specification, all fields of the SampledValue
class except for "value" are optional and may be omitted.

However, the Java class implementation in this library initialized all
fields but "phase" to their default values, and prevented resetting all
but the enum fields to null. Most prominently, the "unit" field was
initialized to "Wh" and could not be reset, preventing the generation of
compliant messages for "Frequency" and "Power.Factor" measurands, for
which the "unit" field needs to be omitted.

Remove the initialization of fields to their default values from the
class constructors, so that they are left as null and will be omitted
when serializing the object.

Exempt null references from the validations in the field setter methods,
so that all fields can be set to null.

Add returning the default value in the field getter methods in case the
stored field value is null, where applicable.

Most importantly, for the "unit" field, only replace null with "Wh" for
"Energy" type measurands. This ensures the null is retained for the
"Frequency" and "Power.Factor" measurands, in compliance with the OCPP
specification.