Fix OCPP V2.0 BootNotificationSpec test
Commit 48743c952d7493f8e7fca88aae8b9bb128cd811f changed the
serialization of ZonedDateTime objects, losing the timezone information.

However, the BootNotification test compares the original object against
the serialized and deserialized object, which no longer matches when the
test machine is configured to a local timezone other than UTC.

Fix the test to take timestamps with UTC timezone so that they can match
ZonedDateTime objects resulting from serialization and deserialization.