Allow listVersion 0 in SendLocalListRequest
The OCPP 1.6 specification mentions no explicit value range for the
listVersion field in the SendLocalListRequest, but specifies that 0
should be returned as the version in case the list is empty. Now there
is no message to clear the local authorization list, so clearing the
list can only be achieved with a SendLocalListRequest with an empty
list, and that is a use case for listVersion 0. In fact the OCPP
certification test tool employs such requests.

Change SendLocalListRequest to allow listVersion 0 rather than only 1 or
higher, and adapt the unit test accordingly.

While at it, bump the version in build.gradle to "1.1" to match the one
from pom.xml.