Allow JSONClient reconfiguration when connected
The WebSocket library allows changing the ping interval immediately
during an ongoing connection. Expose this capability by adding a new
reconfigure() method to the JSONClient class.

To make use of this, set the updated PING_INTERVAL_PARAMETER on the
JSONConfiguration object passed to the JSONClient constructor, then
invoke JSONClient#reconfigure().