Add OCPP 2.0.1 and multiple protocol versions support
Add support for auto-negotiation of the protocol version to use in the
WebSocket handshake, to offer and accept both OCPP 1.6 and OCPP 2.0.1.

To facilitate this, add multiple protocol capable variants of the
JSONClient, JSONServer, SessionFactory, WebSocketListener and
WebSocketTransmitter classes.

Add new IMultiProtocolClientAPI and IMultiProtocolServerAPI interfaces
which extend IClientAPI and IServerAPI, respectively, with the methods
needed for multi-protocol support. Specifically, methods to add feature
profiles (OCPP 1.6) and functions (OCPP 2.0.1 or later) for a specific
protocol version.

The IMultiProtocolClientAPI interface has an additional method to
retrieve the protocol version selected by the server.

The IMultiProtocolServerAPI interface has an additional method to
retrieve the port number the server is listening on, which is useful for
tests where the server is started with port number 0 to let it pick any
available port.

The MultiProtocolJSONClient and MultiProtocolJSONServer classes are
implementations of these two APIs:

- Both take a list of protocol versions to offer/accept as a constructor
  argument, instead of the OCPP 1.6 CoreProfile, which is now optional.
- The client has the new getProtocolVersion() API to retrieve the
  protocol version selected by the server after connecting.
- The server makes the selected protocol version available through the
  getProtocolVersion() API added to the SessionInformation object passed
  to the ServerEvents#newSession() API.

Refactor the existing Client, Server and Session classes and interfaces
to make the Session provide the feature repository to Client and Server,
to allow selecting or configuring it to the protocol version selected in
the WebSocket handshake. Adapt the tests to the removal of the feature
repository from the constructor parameters.

Add the protocol version to the feature repository interface and
implementation.

Add an exemplary unit test for serialization and deserialization of the
OCPP 2.0.1 BootNotificationRequest message. More unit tests should be
added.

Add integration tests for the protocol version auto-negotiation between
client and server.