Add an option for non-blocking OCPP-J connects
Add a new JSONConfiguration parameter "CONNECT_NON_BLOCKING" of Boolean
type, which, if true, causes the OCPP-J clients to make non-blocking
connect() and disconnect() calls to the underlying WebSocket library.

The completion of the connection or disconnection attempt will be
communicated via the callbacks in the ClientEvents object passed to the
OCPP library's connect() method. Note that a failed connection attempt
(connection refused or timeout, if configured) yields a call to the
connectionClosed() callback.