Add getLastError() API to MultiProtocolJSONClient
To allow determining the cause of a disconnection, add a new API to
MultiProtocolJSONClient to retrieve the Exception passed from the
WebSocket library in the onError() callback.

This is required to allow logging the OCPP 2.0.1 Security Events
"InvalidTLSVersion" and "InvalidTLSCipherSuite", by examining the type
of Exception and its message string.