Add OCPP 2.1 support
Add Java classes for all OCPP 2.1 messages and types as well as client
and server event handler interfaces and function classes corresponding
to the OCPP 2.1 "functional blocks."

Add OCPP 2.1 to the ProtocolVersion enum.

Specify UTF-8 encoding for the Java source files in build.gradle and
pom.xml to avoid compiler warnings about the UTF-8 sequences in the
source files.