Implement OCPP 2.1 RPC Layer Extensions
Implement the CALLRESULTERROR RPC as a response to a CALLRESULT, in case
that failed internal validation, or if the completion action last added
to the CompletionStage returned by send() using .whenComplete[Async]()
throws any exception.

Add an optional confirmationError() callback method to the ClientEvents
and ServerEvents interfaces to allow the application to get notified
about an incoming CALLRESULTERROR. The application may use the uniqueId
parameter to match it to the Request#getOcppMessageId() of one of the
requests it last responded to, provided it keeps track of them.

Implement the SEND RPC as requests which do not have a confirmation
type. These are sent through the existing send() method, returning a
CompletableFuture which is already completed when the method returns,
to which a completion action can be added to check whether a local
exception prevented the request from being sent.

Fix the only OCPP message using the SEND RPC, NotifyPeriodicEventStream,
and add its missing Feature and Function handlers. The handler for this
message is a void method, as it has no response.

Remove sending CALLERROR in response to anything other than a CALL,
except for the RpcFrameworkError when the message could not be parsed.

Simplify the pendingPromises synchronization in Session by using the
ConcurrentHashMap class rather than HashMap.

Fix a few typos encountered along the way.